+++ packages/protocol/hardhat.config.ts
@@ -139,7 +139,7 @@ const config: HardhatUserConfig = {
         settings: {
             optimizer: {
                 enabled: true,
-                runs: 10000,
+                runs: 200,
             },
             outputSelection: {
                 "*": {