+++ packages/relayer/message/is_profitable_test.go
@@ -23,7 +23,7 @@ func Test_isProfitable(t *testing.T) {
 		{
 			"zeroProcessingFee",
 			bridge.IBridgeMessage{
-				ProcessingFee: big.NewInt(0),
+				Fee: big.NewInt(0),
 			},
 			big.NewInt(1),
 			false,
@@ -39,8 +39,8 @@ func Test_isProfitable(t *testing.T) {
 		{
 			"lowProcessingFeeHighCost",
 			bridge.IBridgeMessage{
-				ProcessingFee: new(big.Int).Sub(mock.ProcessMessageTx.Cost(), big.NewInt(1)),
-				DestChainId:   big.NewInt(167001),
+				Fee:         new(big.Int).Sub(mock.ProcessMessageTx.Cost(), big.NewInt(1)),
+				DestChainId: big.NewInt(167001),
 			},
 			big.NewInt(1000000),
 			false,
@@ -49,8 +49,8 @@ func Test_isProfitable(t *testing.T) {
 		{
 			"profitableProcessingFee",
 			bridge.IBridgeMessage{
-				ProcessingFee: new(big.Int).Add(mock.ProcessMessageTx.Cost(), big.NewInt(1)),
-				DestChainId:   big.NewInt(167001),
+				Fee:         new(big.Int).Add(mock.ProcessMessageTx.Cost(), big.NewInt(1)),
+				DestChainId: big.NewInt(167001),
 			},
 			big.NewInt(1),
 			true,